simplifying user & password validation
