+++ fileStorage/fileStorage.json
@@ -1,1 +1,1 @@
-{"Doe John":"$2b$10$B8LnMCaZBNNzq5nsSbZSBeFjgQXiQgHJeQyxoHsAJlXhLG4DikFbu","Dani Janos":"$2b$10$Vqf0j2/lIpk5tz6/nz188OeXH8tu.ZyblFRNL6Y75yGUyOEBh9Ula","Pall Arpad":"$2b$10$0rsfIGLh88gN5OiOjJkopugnORkTaCYKNPmI95cptg.GuJs4.LNy6"}
+{"Doe John":"$2b$10$B8LnMCaZBNNzq5nsSbZSBeFjgQXiQgHJeQyxoHsAJlXhLG4DikFbu","Dani Janos":"$2b$10$Vqf0j2/lIpk5tz6/nz188OeXH8tu.ZyblFRNL6Y75yGUyOEBh9Ula","Pall Arpad":"$2b$10$0rsfIGLh88gN5OiOjJkopugnORkTaCYKNPmI95cptg.GuJs4.LNy6","Sandor":"$2b$10$q/HJTuTt6ksJQx24VfSPCeFb0I9A.lPTPfefl9a0jTJfx5UrFh1iK","Sandor Lorant":"$2b$10$UNgaNctLVnm/TVu7imIOUOsmVcvQ79odWsbnsYnpAWs/TYC/JRXXG"}
